+++ src/util/ini.h
@@ -93,7 +93,7 @@ typedef struct FTIT_iniparser {
     char*       (*getString)(struct FTIT_iniparser*, const char*);
     int         (*getInt)(struct FTIT_iniparser*, const char*);
     bool        (*getBool)(struct FTIT_iniparser*, const char*);
-    int         (*getLong)(struct FTIT_iniparser*, const char*);
+    long         (*getLong)(struct FTIT_iniparser*, const char*);
     int         (*set)(struct FTIT_iniparser*, const char*,
                                  const char*);
     int         (*dump)(struct FTIT_iniparser*);
@@ -200,7 +200,7 @@ bool FTI_IniparserGetBool(FTIT_iniparser*, const char* key);
  
 
 --------------------------------------------------------------------------**/
-int FTI_IniparserGetLong(FTIT_iniparser*, const char* key);
+long FTI_IniparserGetLong(FTIT_iniparser*, const char* key);
 
 /**--------------------------------------------------------------------------
